Add new StateManager, State and Transition
In order to facilitate easier handling of game states, I've created a
new StateManager struct, inspired by the way Amethyst handles states and
their transitions. StateManager impls EventHandler, and is therefore the
main struct passed to ggez. You can then pass it anything that impls the
new State trait, which is essentially EventHandler but returns
Tranistions so it can be used by the StateManager. Note that states
currently can't pass data between each other; not sure what the best way
to do that without fighting the borrow checker is.